Refactored auth-service controller payloads: - deleted TokenPairResponsePayload - moved RegisterResponsePayload to user-manager controller payloads and removed validation - moved UserRegistrationRequestPayload to user-manager controller payloads
