+++ user-manager-service/src/main/java/com/crmfoodestablishment/userauthservice/authservice/controller/payload/LoginRequestPayload.java
@@ -1,4 +1,4 @@
-package com.crmfoodestablishment.user_auth_service.auth_sevice.controller.payload;
+package com.crmfoodestablishment.userauthservice.authservice.controller.payload;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
