+++ user-manager-service/src/main/java/com/crmfoodestablishment/userauthservice/usermanager/controller/payload/RegisterResponsePayload.java
@@ -1,0 +1,16 @@
+package com.crmfoodestablishment.userauthservice.usermanager.controller.payload;
+
+import com.crmfoodestablishment.userauthservice.authservice.token.TokenPair;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+@AllArgsConstructor
+@Getter
+@Setter
+public class RegisterResponsePayload {
+
+    private String urlOfCreatedUser;
+
+    private TokenPair tokenPair;
+}
